整理分组信息，

+++ model/BPHCenter/src/main/java/com/tianyi/bph/web/controller/dutydata/PoliceGroupController.java
@@ -1,11 +1,14 @@
 package com.tianyi.bph.web.controller.dutydata;
 
+import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.Map;
 import java.util.List;
 import javax.servlet.http.HttpServletRequest;
 
+import net.sf.json.JSONArray;
 import net.sf.json.JSONObject;
+import net.sf.json.JsonConfig;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
@@ -15,6 +18,10 @@ import org.springframework.web.bind.annotation.ResponseBody;
 
 import com.tianyi.bph.common.MessageCode;
 import com.tianyi.bph.common.ReturnResult;
+import com.tianyi.bph.domain.duty.PoliceGroup;
+import com.tianyi.bph.domain.duty.PoliceGroupMember;
+import com.tianyi.bph.domain.duty.VehicleGroup;
+import com.tianyi.bph.domain.duty.VehicleGroupMember;
 import com.tianyi.bph.query.duty.PoliceGroupMemberVM;
 import com.tianyi.bph.query.duty.PoliceGroupVM;
 import com.tianyi.bph.service.duty.PoliceGroupService;
@@ -28,7 +35,6 @@ public class PoliceGroupController {
 	protected PoliceGroupService policeGroupService;
 	@Autowired
 	private OrganService organService;
-	
 
 	/**
 	 * 获取警员分组列表
@@ -43,31 +49,60 @@ public class PoliceGroupController {
 	public @ResponseBody
 	ReturnResult List(
 			@RequestParam(value = "policeGroup_Query", required = false) String query,
-			@RequestParam(value = "page", required = false) Integer page,
-			@RequestParam(value = "rows", required = false) Integer rows,
 			HttpServletRequest request) {
 
 		JSONObject joQuery = JSONObject.fromObject(query);
 
 		int orgId = joQuery.getInt("orgId");
-		String orgCode = joQuery.getString("orgCode");
+		int page = joQuery.getInt("page");
+		int pageSize = joQuery.getInt("pageSize");
 
 		Map<String, Object> map = new HashMap<String, Object>();
 		page = page == 0 ? 1 : page;
-		map.put("pageStart", (page - 1) * rows);
-		map.put("pageSize", rows);
-		map.put("orgCode", orgCode);
+		map.put("pageStart", (page - 1) * pageSize);
+		map.put("pageSize", pageSize);
 		map.put("orgId", orgId);
-
-		map.put("inSubOrg", 0);
 
 		int total = policeGroupService.loadVMCountByOrgId(map);
 
 		List<PoliceGroupVM> pgvms = policeGroupService.loadVMListByOrgId(map);
 
+		return ReturnResult
+				.MESSAGE(MessageCode.STATUS_SUCESS, "", total, pgvms);
 
-		return ReturnResult.MESSAGE(MessageCode.STATUS_SUCESS,"", total, pgvms);
+	}
 
+	/**
+	 * 保存车辆分组信息
+	 * 
+	 * @param vehicleGroup
+	 * @param request
+	 * @return
+	 */
+	@RequestMapping(value = "savePoliceGroup.do")
+	public @ResponseBody
+	ReturnResult savePoliceGroup(
+			@RequestParam(value = "policeGroup", required = false) String policeGroup,
+			HttpServletRequest request) {
+		try {
+			JSONObject joPG = JSONObject.fromObject(policeGroup);
+
+			PoliceGroup pg = new PoliceGroup();
+			pg.setId(joPG.getInt("id"));
+			pg.setName(joPG.getString("name"));
+			pg.setShareType(joPG.getInt("shareType"));
+			pg.setOrgId(joPG.getInt("orgId"));
+			pg.setPlatformId(1);
+			pg.setSyncState(true);
+			JSONArray jaOrgIds = joPG.getJSONArray("shareOrgIds");
+
+			Object[] orgIds = jaOrgIds.toArray();
+
+			policeGroupService.savePoliceGroup(pg, orgIds);
+			return ReturnResult.MESSAGE(MessageCode.STATUS_SUCESS, "", 0, null);
+		} catch (Exception ex) {
+			return ReturnResult.MESSAGE(MessageCode.STATUS_FAIL, "", 0, null);
+		}
 	}
 
 	/**
@@ -83,10 +118,8 @@ public class PoliceGroupController {
 	public @ResponseBody
 	ReturnResult loadMemberByGroupId(
 			@RequestParam(value = "groupId", required = false) Integer groupId,
-			@RequestParam(value = "page", required = false) Integer page,
-			@RequestParam(value = "rows", required = false) Integer rows,
 			HttpServletRequest request) {
-   
+
 		Map<String, Object> map = new HashMap<String, Object>();
 		map.put("pageStart", 0);
 		map.put("pageSize", 10);
@@ -96,11 +129,44 @@ public class PoliceGroupController {
 
 		List<PoliceGroupMemberVM> ms = policeGroupService
 				.loadMemberByGroupId(map);
-  
 
 		return ReturnResult.MESSAGE(MessageCode.STATUS_SUCESS, "", total, ms);
 	}
-	
+
+	/**
+	 * 增加组成员
+	 * 
+	 * @param members
+	 * @param request
+	 * @return
+	 */
+	@RequestMapping(value = "appendMember.do")
+	public @ResponseBody
+	ReturnResult appendMember(
+			@RequestParam(value = "members", required = false) String members,
+			HttpServletRequest request) {
+		try {
+			JSONArray jaMembers = JSONArray.fromObject(members); 
+			List<?> list2 = JSONArray.toList(jaMembers,
+					new PoliceGroupMember(), new JsonConfig());// 参数1为要转换的JSONArray数据，参数2为要转换的目标数据，即List盛装的数据
+
+			List<PoliceGroupMember> ls = new ArrayList<PoliceGroupMember>();
+
+			for (Object o : list2) {
+				PoliceGroupMember pgm = (PoliceGroupMember) o;
+				ls.add(pgm);
+			}
+			policeGroupService.appendMemeber(ls);
+			return ReturnResult.MESSAGE(MessageCode.STATUS_SUCESS, "删除成功", 0,
+					null);
+
+		} catch (Exception ex) {
+			return ReturnResult.MESSAGE(MessageCode.STATUS_FAIL, "查询数据出错", 0,
+					null);
+		}
+
+	}
+
 	/**
 	 * 删除组成员
 	 * 
@@ -110,16 +176,22 @@ public class PoliceGroupController {
 	 */
 	@RequestMapping(value = "delMemberById.do")
 	public @ResponseBody
-	String delMemberById(
+	ReturnResult delMemberById(
 			@RequestParam(value = "memberId", required = false) Integer memberId,
 			HttpServletRequest request) {
-
-		policeGroupService.delMemberById(memberId);
-
-		JSONObject rs = new JSONObject();
-		rs.accumulate("isSuccess", true);
-
-		return rs.toString();
+		try {
+			if (memberId > 0) {
+				policeGroupService.delMemberById(memberId);
+				return ReturnResult.MESSAGE(MessageCode.STATUS_SUCESS, "删除成功",
+						0, null);
+			} else {
+				return ReturnResult.MESSAGE(MessageCode.STATUS_FAIL, "删除成功", 0,
+						null);
+			}
+		} catch (Exception ex) {
+			return ReturnResult.MESSAGE(MessageCode.STATUS_FAIL, "查询数据出错", 0,
+					null);
+		}
 	}
 
 	/**
@@ -131,36 +203,83 @@ public class PoliceGroupController {
 	 */
 	@RequestMapping(value = "cleanMemberByGroupId.do")
 	public @ResponseBody
-	String cleanMemberByGroupId(
+	ReturnResult cleanMemberByGroupId(
 			@RequestParam(value = "policeGroupId", required = false) Integer policeGroupId,
 			HttpServletRequest request) {
-
-		policeGroupService.cleanMember(policeGroupId);
-
-		return "";
+		try {
+			if (policeGroupId > 0) {
+				policeGroupService.cleanMember(policeGroupId);
+				return ReturnResult.MESSAGE(MessageCode.STATUS_SUCESS, "删除成功",
+						0, null);
+			} else {
+				return ReturnResult.MESSAGE(MessageCode.STATUS_FAIL, "删除成功", 0,
+						null);
+			}
+		} catch (Exception ex) {
+			return ReturnResult.MESSAGE(MessageCode.STATUS_FAIL, "查询数据出错", 0,
+					null);
+		}
 	}
-	
+
 	/**
-	 * 删除警员组
+	 * 删除车辆组
 	 * 
-	 * @param policeGroupId
+	 * @param vehicleGroupId
 	 * @param request
 	 * @return
 	 */
 	@RequestMapping(value = "deletePoliceGroup.do")
 	public @ResponseBody
-	String deletePoliceGroup(
+	ReturnResult deletePoliceGroup(
 			@RequestParam(value = "policeGroupId", required = false) Integer policeGroupId,
 			HttpServletRequest request) {
-
-		policeGroupService.deleteById(policeGroupId);
-
-		JSONObject rs = new JSONObject();
-		rs.accumulate("isSuccess", true);
-		rs.accumulate("msg", "");
-
-		return rs.toString();
+		try {
+			if (policeGroupId > 0) {
+				policeGroupService.deleteById(policeGroupId);
+				return ReturnResult.MESSAGE(MessageCode.STATUS_SUCESS, "删除成功",
+						0, null);
+			} else {
+				return ReturnResult.MESSAGE(MessageCode.STATUS_FAIL, "删除成功", 0,
+						null);
+			}
+		} catch (Exception ex) {
+			return ReturnResult.MESSAGE(MessageCode.STATUS_FAIL, "查询数据出错", 0,
+					null);
+		}
 	}
 
+	/**
+	 * 判断是否有有分组存在
+	 * 
+	 * 判断是否分组名称重复；
+	 */
+	@RequestMapping(value = "isExistGroup.do")
+	public @ResponseBody
+	ReturnResult isExistGroup(
+			@RequestParam(value = "name", required = false) String name,
+			@RequestParam(value = "optType", required = false) Integer optType,
+			@RequestParam(value = "groupId", required = false) Integer groupId,
+			@RequestParam(value = "orgId", required = false) Integer orgId)
+			throws Exception {
+		try {
+			Map<String, Object> map = new HashMap<String, Object>();
+			map.put("name", name);
+			map.put("groupId", groupId);
+			map.put("optType", optType);
+			map.put("orgId", orgId);
+			List<PoliceGroup> vehicleGroup = policeGroupService
+					.findByNameAndOrg(map);
+			if (vehicleGroup.size() > 0) {
+				return ReturnResult.MESSAGE(MessageCode.STATUS_FAIL, "", 0,
+						null);
+			} else {
+				return ReturnResult.MESSAGE(MessageCode.STATUS_SUCESS, "", 0,
+						null);
+			}
+		} catch (Exception ex) {
+			return ReturnResult.MESSAGE(MessageCode.STATUS_FAIL, "查询数据出错", 0,
+					null);
+		}
+	}
 
 }
